QCamera2: Properly fix ARRAY_SIZE() comparison warning
Array indices are never negative, so it doesn't make sense to cast the
ARRAY_SIZE() value as a signed integer.

+++ camera/QCamera2/HAL/QCamera2Capabilities.h
@@ -27,7 +27,7 @@
  *
  */
 
-#define ARRAY_SIZE(a) (int)(sizeof(a)/sizeof((a)[0]))
+#define ARRAY_SIZE(a) (sizeof(a)/sizeof((a)[0]))
 
 static const cam_dimension_t new_pic_sizes_cam0[] = {
     {4208, 3120},
